% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMT.R
\name{boosting_glm}
\alias{boosting_glm}
\title{Multiple Testing Boosting model selection procedure}
\usage{
boosting_glm(
  y,
  X,
  Xcond = NULL,
  pval = 0.05,
  delta1 = 1,
  delta2 = 1,
  link = c("logit", "probit", "mean"),
  maxit = 50
)
}
\arguments{
\item{y}{Numeric vector (length T), typically 0/1.}

\item{X}{Numeric matrix (T x N) of candidate regressors.}

\item{Xcond}{Optional numeric matrix (T x Kc) of always-included controls.}

\item{pval}{Numeric scalar in (0,1).}

\item{delta1}{Numeric scalar >= 1. Default 1.}

\item{delta2}{Numeric scalar >= 1. Default 1.}

\item{link}{One of "logit", "probit", "mean". Default "logit".}

\item{maxit}{Max iterations for GLM fitting.}
}
\value{
Logical vector of length N indicating selected regressors.
}
\description{
Multiple Testing Boosting model selection procedure
}
